Camera Controller Done for now
Finished the camera controller for now, it is going to need a lot more work though. For the time being I wanted to move on to getting the other mouse controls started, this will be the last code commit today though. Going to spend the rest of the day to make the github page look like something xD.